+++ Changes.md
@@ -2,6 +2,7 @@ Version Information
 ===================
 Version 3.6.0.3
   1. Add H5P custom CSS setting from Essential.
+  2. Add scrollable drawer block region.
 
 Version 3.6.0.2
   1. Moodle 3.6.1 version.